+++ modules/kravis/src/main/kotlin/org/kalasim/plot/kravis/TimeDiscretizer.kt
@@ -1,8 +1,8 @@
 package org.kalasim.plot.kravis
 
-import kotlinx.datetime.Instant
 import org.kalasim.SimTime
 import kotlin.time.DurationUnit
+import kotlin.time.Instant
 
 fun interface TimeDiscretizer {
     fun discretize(time: SimTime): SimTime