+++ src/main/kotlin/org/kalasim/Component.kt
@@ -20,6 +20,7 @@ import kotlin.math.min
 import kotlin.reflect.*
 import kotlin.time.Duration
 import kotlin.time.DurationUnit
+import kotlin.time.isDistantFuture
 
 
 internal const val EPS = 1E-8