+++ src/main/kotlin/org/kalasim/analysis/Events.kt
@@ -6,6 +6,7 @@ import org.json.JSONObject
 import org.kalasim.*
 import org.kalasim.misc.*
 import kotlin.math.absoluteValue
+import kotlin.time.isDistantFuture
 
 enum class ResourceEventType { REQUESTED, CLAIMED, RELEASED, PUT, TAKE }