+++ src/main/kotlin/org/kalasim/misc/TestUtil.kt
@@ -1,8 +1,8 @@
 package org.kalasim.misc
 
-import kotlinx.datetime.Instant
 import org.kalasim.*
 import kotlin.time.DurationUnit
+import kotlin.time.Instant
 
 class TestUtil {
 
@@ -14,7 +14,7 @@ class TestUtil {
 }
 
 fun createTestSimulation(
-    startDate: Instant = Instant.fromEpochMilliseconds(0),
+    startDate: SimTime = Instant.fromEpochMilliseconds(0),
     enableComponentLogger: Boolean = true,
     /** The duration unit of this environment. Every tick corresponds to a unit duration. See https://www.kalasim.org/basics/#running-a-simulation */
     tickDurationUnit: DurationUnit = DurationUnit.MINUTES,